rootfs-postcommands: Cope with empty IMAGE_LINK_NAME in write_image_manifest
Ensure that we don't create a symlink named ".manifest" if IMAGE_LINK_NAME
is empty.

+++ meta/classes/rootfs-postcommands.bbclass
@@ -260,7 +260,7 @@ python write_image_manifest () {
     with open(manifest_name, 'w+') as image_manifest:
         image_manifest.write(format_pkg_list(pkgs, "ver"))
 
-    if os.path.exists(manifest_name):
+    if os.path.exists(manifest_name) and link_name:
         manifest_link = deploy_dir + "/" + link_name + ".manifest"
         if os.path.lexists(manifest_link):
             os.remove(manifest_link)
